Moved constants and logo ascii art out of the main file
