+++ archey/archey.py
@@ -5,6 +5,7 @@ import json
 import os
 import re
 import sys
+
 from enum import Enum
 from glob import glob
 from subprocess import CalledProcessError, DEVNULL, PIPE, Popen, \
@@ -12,392 +13,13 @@ from subprocess import CalledProcessError, DEVNULL, PIPE, Popen, \
 
 import distro
 
-# ----- Distributions fingerprints ---- #
-
-class Distributions(Enum):
-    ARCH_LINUX = 'Arch.*'
-    BUNSENLABS = 'BunsenLabs'
-    CRUNCHBANG = 'CrunchBang'
-    DEBIAN = '(Rasp|De)bian'
-    FEDORA = 'Fedora'
-    GENTOO = 'Gentoo'
-    KALI_LINUX = 'Kali'
-    MANJARO_LINUX = 'Manjaro ?Linux'
-    LINUX = 'Linux'
-    LINUX_MINT = 'Linux ?Mint'
-    OPENSUSE = 'openSUSE'
-    RED_HAT = '(Red ?Hat|RHEL)'
-    SLACKWARE = 'Slackware'
-    UBUNTU = 'Ubuntu'
-    WINDOWS = 'Windows'
-
-
-# ------------ Dictionaries ----------- #
-
-COLOR_DICT = {
-    Distributions.ARCH_LINUX: ['\x1b[0;34m', '\x1b[1;34m'],
-    Distributions.BUNSENLABS: ['\x1b[1;37m', '\x1b[1;33m', '\x1b[0;33m'],
-    Distributions.CRUNCHBANG: ['\x1b[1;37m', '\x1b[1;37m'],
-    Distributions.DEBIAN: ['\x1b[0;31m', '\x1b[1;31m'],
-    Distributions.FEDORA: ['\x1b[1;37m', '\x1b[0;34m'],
-    Distributions.GENTOO: ['\x1b[1;37m', '\x1b[1;35m'],
-    Distributions.KALI_LINUX: ['\x1b[1;37m', '\x1b[1;34m'],
-    Distributions.MANJARO_LINUX: ['\x1b[0;32m', '\x1b[1;32m'],
-    Distributions.LINUX: ['\x1b[1;33m', '\x1b[1;37m'],
-    Distributions.LINUX_MINT: ['\x1b[1;37m', '\x1b[1;32m'],
-    Distributions.OPENSUSE: ['\x1b[1;37m', '\x1b[1;32m'],
-    Distributions.RED_HAT: ['\x1b[1;37m', '\x1b[1;31m', '\x1b[0;31m'],
-    Distributions.SLACKWARE: ['\x1b[0;34m', '\x1b[1;34m', '\x1b[1;0m'],
-    Distributions.UBUNTU: ['\x1b[0;31m', '\x1b[1;31m', '\x1b[0;33m'],
-    Distributions.WINDOWS: ['\x1b[1;31m', '\x1b[1;34m',
-                            '\x1b[1;32m', '\x1b[0;33m'],
-    'sensors': ['\x1b[0;32m', '\x1b[0;33m', '\x1b[0;31m'],
-    'clear': '\x1b[0m'
-}
-
-DE_DICT = {
-    'cinnamon': 'Cinnamon',
-    'gnome-session': 'GNOME',
-    'gnome-shell': 'GNOME',
-    'mate-session': 'MATE',
-    'ksmserver': 'KDE',
-    'xfce4-session': 'Xfce',
-    'fur-box-session': 'Fur Box',
-    'lxsession': 'LXDE',
-    'lxqt-session': 'LXQt'
-}
-
-WM_DICT = {
-    'awesome': 'Awesome',
-    'beryl': 'Beryl',
-    'blackbox': 'Blackbox',
-    'bspwm': 'bspwm',
-    'cinnamon': 'Cinnamon',
-    'compiz': 'Compiz',
-    'dwm': 'DWM',
-    'enlightenment': 'Enlightenment',
-    'herbstluftwm': 'herbstluftwm',
-    'fluxbox': 'Fluxbox',
-    'fvwm': 'FVWM',
-    'i3': 'i3',
-    'icewm': 'IceWM',
-    'kwin_x11': 'KWin',
-    'metacity': 'Metacity',
-    'musca': 'Musca',
-    'openbox': 'Openbox',
-    'pekwm': 'PekWM',
-    'qtile': 'QTile',
-    'ratpoison': 'ratpoison',
-    'scrotwm': 'ScrotWM',
-    'stumpwm': 'StumpWM',
-    'subtle': 'Subtle',
-    'monsterwm': 'MonsterWM',
-    'wingo': 'Wingo',
-    'wmaker': 'Window Maker',
-    'wmfs': 'Wmfs',
-    'wmii': 'wmii',
-    'xfwm4': 'Xfwm',
-    'xmonad': 'xmonad'
-}
-
-LOGOS_DICT = {
-    Distributions.ARCH_LINUX: """
-{c[1]}               +                 {r[0]}
-{c[1]}               #                 {r[1]}
-{c[1]}              ###                {r[2]}
-{c[1]}             #####               {r[3]}
-{c[1]}             ######              {r[4]}
-{c[1]}            ; #####;             {r[5]}
-{c[1]}           +##.#####             {r[6]}
-{c[1]}          +##########            {r[7]}
-{c[1]}         ######{c[0]}#####{c[1]}##;          {r[8]}
-{c[1]}        ###{c[0]}############{c[1]}+         {r[9]}
-{c[1]}       #{c[0]}######   #######{c[1]}         {r[10]}
-{c[1]}     {c[0]}.######;     ;###;`\".{c[1]}       {r[11]}
-{c[1]}    {c[0]}.#######;     ;#####.{c[1]}        {r[12]}
-{c[1]}    {c[0]}#########.   .########`{c[1]}      {r[13]}
-{c[1]}   {c[0]}######'           '######{c[1]}     {r[14]}
-{c[1]}  {c[0]};####                 ####;{c[1]}    {r[15]}
-{c[1]}  {c[0]}##'                     '##{c[1]}    {r[16]}
-{c[1]} {c[0]}#'                         `#{c[1]}   {r[17]}\n""",
-    Distributions.BUNSENLABS: """
-{c[0]}           .{c[1]}..{c[0]}+hhy+-`         \0
-{c[0]}        `+hd{c[1]}-{c[0]}+dddd{c[2]}hyso{c[0]}+//:    \0
-{c[0]}      `+dddd{c[1]}:-{c[0]}sdh/.           {r[0]}
-{c[0]}     -hdddddh{c[1]}-.{c[2]}/:{c[0]}             {r[1]}
-{c[0]}    /ddddddddd{c[1]}:```{c[0]}            {r[2]}
-{c[0]}   :ddddddddddd/              {r[3]}
-{c[0]}  `hdddddddddddd+             {r[4]}
-{c[0]}  /dddddddddddddd:            {r[5]}
-{c[0]}  odddds..sddddddh            {r[6]}
-{c[0]}  oddd/    /dddddd:           {r[7]}
-{c[0]}  +dd+      +ddddd+           {r[8]}
-{c[0]}  .dd`      `ddddd+           {r[9]}
-{c[0]}   oh        ydddd:           {r[10]}
-{c[0]}   `o        sdddh`           {r[11]}
-{c[0]}             yddd:            {r[12]}
-{c[0]}            `dddo             {r[13]}
-{c[0]}     :s     :dds              {r[14]}
-{c[0]}     yd/    yd+               {r[15]}
-{c[0]}   `sddy   :h-                {r[16]}
-{c[0]}  `sddys`  :`                 {r[17]}
-{c[0]} -hdy+`y+yo./+/:-             \0
-{c[0]} ...  .o++oso+/               \n""",
-    Distributions.CRUNCHBANG: """
-{c[0]}                 ___       ___      _    {r[0]}
-{c[0]}                /  /      /  /     | |   {r[1]}
-{c[0]}               /  /      /  /      | |   {r[2]}
-{c[0]}              /  /      /  /       | |   {r[3]}
-{c[0]}      _______/  /______/  /______  | |   {r[4]}
-{c[0]}     /______   _______   _______/  | |   {r[5]}
-{c[0]}           /  /      /  /          | |   {r[6]}
-{c[0]}          /  /      /  /           | |   {r[7]}
-{c[0]}         /  /      /  /            | |   {r[8]}
-{c[0]}  ______/  /______/  /______       | |   {r[9]}
-{c[0]} /_____   _______   _______/       | |   {r[10]}
-{c[0]}      /  /      /  /               | |   {r[11]}
-{c[0]}     /  /      /  /                |_|   {r[12]}
-{c[0]}    /  /      /  /                  _    {r[13]}
-{c[0]}   /  /      /  /                  | |   {r[14]}
-{c[0]}  /__/      /__/                   |_|   {r[15]}
-{c[0]}                                         {r[16]}
-{c[0]}                                         {r[17]}\n""",
-    Distributions.DEBIAN: """
-{c[1]}                                  {r[0]}
-{c[1]}          _sudZUZ#Z#XZo=_         {r[1]}
-{c[1]}       _jmZZ2!!~---~!!X##wx       {r[2]}
-{c[1]}    .<wdP~~            -!YZL,     {r[3]}
-{c[1]}   .mX2'       _xaaa__     XZ[.   {r[4]}
-{c[1]}   oZ[      _jdXY!~?S#wa   ]Xb;   {r[5]}
-{c[1]}  _#e'     .]X2(     ~Xw|  )XXc   {r[6]}
-{c[1]} .2Z`      ]X[.       xY|  ]oZ(   {r[7]}
-{c[1]} .2#;      )3k;     _s!~   jXf`   {r[8]}
-{c[1]}  {c[0]}1Z>{c[1]}      {c[0]}-]Xb/{c[1]}    {c[0]}~{c[1]}    {c[0]}__#2({c[1]}    {r[9]}
-{c[1]}  {c[0]}-Zo;{c[1]}       {c[0]}+!4ZwerfgnZZXY'{c[1]}      {r[10]}
-{c[1]}   {c[0]}*#[,{c[1]}        {c[0]}~-?!!!!!!-~{c[1]}        {r[11]}
-{c[1]}    {c[0]}XUb;.{c[1]}                         {r[12]}
-{c[1]}     {c[0]})YXL,,{c[1]}                       {r[13]}
-{c[1]}       {c[0]}+3#bc,{c[1]}                     {r[14]}
-{c[1]}         {c[0]}-)SSL,,{c[1]}                  {r[15]}
-{c[1]}            {c[0]}~~~~~{c[1]}                 {r[16]}
-{c[1]}                                  {r[17]}\n""",
-    Distributions.FEDORA: """
-{c[1]}           :/------------://          {r[0]}
-{c[1]}        :------------------://        {r[1]}
-{c[1]}      :-----------{c[0]}/shhdhyo/{c[1]}-://       {r[2]}
-{c[1]}    /-----------{c[0]}omMMMNNNMMMd/{c[1]}-:/      {r[3]}
-{c[1]}   :-----------{c[0]}sMMMdo:/{c[1]}       -:/     {r[4]}
-{c[1]}  :-----------{c[0]}:MMMd{c[1]}-------    --:/    {r[5]}
-{c[1]}  /-----------{c[0]}:MMMy{c[1]}-------    ---/    {r[6]}
-{c[1]} :------    --{c[0]}/+MMMh/{c[1]}--        ---:   {r[7]}
-{c[1]} :---     {c[0]}oNMMMMMMMMMNho{c[1]}     -----:   {r[8]}
-{c[1]} :--      {c[0]}+shhhMMMmhhy++{c[1]}   ------:    {r[9]}
-{c[1]} :-      -----{c[0]}:MMMy{c[1]}--------------/    {r[10]}
-{c[1]} :-     ------{c[0]}/MMMy{c[1]}-------------:     {r[11]}
-{c[1]} :-      ----{c[0]}/hMMM+{c[1]}------------:      {r[12]}
-{c[1]} :--{c[0]}:dMMNdhhdNMMNo{c[1]}-----------:        {r[13]}
-{c[1]} :---{c[0]}:sdNMMMMNds:{c[1]}----------:          {r[14]}
-{c[1]} :------{c[0]}:://:{c[1]}-----------://           {r[15]}
-{c[1]} :--------------------://             {r[16]}
-{c[1]}                                      {r[17]}\n""",
-    Distributions.GENTOO: """
-{c[1]}          -/oyddmdhs+:.                {r[0]}
-{c[1]}      -o{c[0]}dNMMMMMMMMNNmhy+{c[1]}-`             {r[1]}
-{c[1]}    -y{c[0]}NMMMMMMMMMMMNNNmmdhy{c[1]}+-           {r[2]}
-{c[1]}  `o{c[0]}mMMMMMMMMMMMMNmdmmmmddhhy{c[1]}/`        {r[3]}
-{c[1]}  om{c[0]}MMMMMMMMMMMN{c[1]}hhyyyo{c[0]}hmdddhhhd{c[1]}o`      {r[4]}
-{c[1]} .y{c[0]}dMMMMMMMMMMd{c[1]}hs++so/s{c[0]}mdddhhhhdm{c[1]}+`    {r[5]}
-{c[1]}  oy{c[0]}hdmNMMMMMMMN{c[1]}dyooy{c[0]}dmddddhhhhyhN{c[1]}d.   {r[6]}
-{c[1]}   :o{c[0]}yhhdNNMMMMMMMNNNmmdddhhhhhyym{c[1]}Mh   {r[7]}
-{c[1]}     .:{c[0]}+sydNMMMMMNNNmmmdddhhhhhhmM{c[1]}my   {r[8]}
-{c[1]}        /m{c[0]}MMMMMMNNNmmmdddhhhhhmMNh{c[1]}s:   {r[9]}
-{c[1]}     `o{c[0]}NMMMMMMMNNNmmmddddhhdmMNhs{c[1]}+`    {r[10]}
-{c[1]}   `s{c[0]}NMMMMMMMMNNNmmmdddddmNMmhs{c[1]}/.      {r[11]}
-{c[1]}  /N{c[0]}MMMMMMMMNNNNmmmdddmNMNdso{c[1]}:`        {r[12]}
-{c[1]} +M{c[0]}MMMMMMNNNNNmmmmdmNMNdso{c[1]}/-           {r[13]}
-{c[1]} yM{c[0]}MNNNNNNNmmmmmNNMmhs+/{c[1]}-`             {r[14]}
-{c[1]} /h{c[0]}MMNNNNNNNNMNdhs++/{c[1]}-`                {r[15]}
-{c[1]} `/{c[0]}ohdmmddhys+++/:{c[1]}.`                   {r[16]}
-{c[1]} `-//////:--.                          {r[17]}\n""",
-    Distributions.KALI_LINUX: """
-{c[0]}      ,.....                                        \0
-{c[0]}  ----`     `..,;:ccc,.                             \0
-{c[0]}           ......''';lxO.                           {r[0]}
-{c[0]} .....''''..........,:ld;                           {r[1]}
-{c[0]}            .';;;:::;,,.x,                          {r[2]}
-{c[0]}       ..'''.            0Xxoc:,.  ...              {r[3]}
-{c[0]}   ....                ,ONkc;,;cokOdc',.            {r[4]}
-{c[0]}  .                   OMo           ':{c[1]}d{c[0]}o.           {r[5]}
-{c[0]}                     dMc               :OO;         {r[6]}
-{c[0]}                     0M.                 .:o.       {r[7]}
-{c[0]}                     ;Wd                            {r[8]}
-{c[0]}                      ;XO,                          {r[9]}
-{c[0]}                        ,d0Odlc;,..                 {r[10]}
-{c[0]}                            ..',;:cdOOd::,.         {r[11]}
-{c[0]}                                     .:d;.':;.      {r[12]}
-{c[0]}                                        'd,  .'     {r[13]}
-{c[0]}                                          ;l   ..   {r[14]}
-{c[0]}                                           .o       {r[15]}
-{c[0]}                                             c      {r[16]}
-{c[0]}                                             .'     {r[17]}
-{c[0]}                                              .     \n""",
-    Distributions.MANJARO_LINUX: """
-{c[0]} $$$$$$$$$$$$$$$$  $$$$$$$   {r[0]}
-{c[0]} M77777777777777M  M77777M   {r[1]}
-{c[0]} M77777777777777M  M77777M   {r[2]}
-{c[0]} M77777MMMMMMMMMM  M77777M   {r[3]}
-{c[0]} M77777M           M77777M   {r[4]}
-{c[0]} M77777M  $$$$$$$  M77777M   {r[5]}
-{c[0]} MMMMMMM  M77777M  M77777M   {r[6]}
-{c[0]}          M77777M  M77777M   {r[7]}
-{c[0]} $$$$$$$  M77777M  M77777M   {r[8]}
-{c[0]} M77777M  M77777M  M77777M   {r[9]}
-{c[0]} M77777M  M77777M  M77777M   {r[10]}
-{c[0]} M77777M  M77777M  M77777M   {r[11]}
-{c[0]} M77777M  M77777M  M77777M   {r[12]}
-{c[0]} M77777M  M77777M  M77777M   {r[13]}
-{c[0]} M77777M  M77777M  M77777M   {r[14]}
-{c[0]} M77777M  M77777M  M77777M   {r[15]}
-{c[0]} M77777M  M77777M  M77777M   {r[16]}
-{c[0]} MMMMMMM  MMMMMMM  MMMMMMM   {r[17]}\n""",
-    Distributions.LINUX: """
-{c[1]}                           {r[0]}
-{c[1]}          a8888b.          {r[1]}
-{c[1]}         d888888b.         {r[2]}
-{c[1]}         8P"YP"Y88         {r[3]}
-{c[1]}         8|o||o|88         {r[4]}
-{c[1]}         8{c[0]}\\vvvv/{c[1]}88         {r[5]}
-{c[1]}         8{c[0]} \\vv/ {c[1]}Y8.        {r[6]}
-{c[1]}        d/  {c[0]}`'{c[1]}  \\8b.       {r[7]}
-{c[1]}      .dP   .     Y8b.     {r[8]}
-{c[1]}     d8:'   "   `::88b.    {r[9]}
-{c[1]}    d8"           `Y88b    {r[10]}
-{c[1]}   :8P     '       :888    {r[11]}
-{c[1]}    8a.    :      _a88P    {r[12]}
-{c[1]}  {c[0]}._/"{c[1]}Yaa_ :    .{c[0]}| {c[1]}88P{c[0]}|{c[1]}    {r[13]}
-{c[1]} {c[0]}\\++++{c[1]}YP"      `{c[0]}| {c[1]}8P{c[0]}++\\.{c[1]}   {r[14]}
-{c[1]} {c[0]}/+++++\\.{c[1]}_____.d{c[0]}|+++++/{c[1]}    {r[15]}
-{c[1]}  {c[0]}\\++++++){c[1]}888888P{c[0]}\\+++/{c[1]}     {r[16]}
-{c[1]}                           {r[17]}\n""",
-    Distributions.LINUX_MINT: """
-{c[1]}                                       {r[0]}
-{c[1]} MMMMMMMMMMMMMMMMMMMMMMMMMmds+.        {r[1]}
-{c[1]} MMm----::-://////////////oymNMd+`     {r[2]}
-{c[1]} MMd      {c[0]}/++{c[1]}                -sNMd:    {r[3]}
-{c[1]} MMNso/`  {c[0]}dMM{c[1]}    {c[0]}`.::-. .-::.`{c[1]} .hMN:   {r[4]}
-{c[1]} ddddMMh  {c[0]}dMM{c[1]}   {c[0]}:hNMNMNhNMNMNh:`{c[1]} NMm   {r[5]}
-{c[1]}     NMm  {c[0]}dMM{c[1]}  {c[0]}.NMN/-+MMM+-/NMN`{c[1]} dMM   {r[6]}
-{c[1]}     NMm  {c[0]}dMM{c[1]}  {c[0]}-MMm{c[1]}  {c[0]}`MMM{c[1]}   {c[0]}dMM.{c[1]} dMM   {r[7]}
-{c[1]}     NMm  {c[0]}dMM{c[1]}  {c[0]}-MMm{c[1]}  {c[0]}`MMM{c[1]}   {c[0]}dMM.{c[1]} dMM   {r[8]}
-{c[1]}     NMm  {c[0]}dMM{c[1]}  {c[0]}.mmd{c[1]}  {c[0]}`mmm{c[1]}   {c[0]}yMM.{c[1]} dMM   {r[9]}
-{c[1]}     NMm  {c[0]}dMM`{c[1]}  {c[0]}..`{c[1]}  {c[0]}`...{c[1]}   {c[0]}ydm.{c[1]} dMM   {r[10]}
-{c[1]}     hMM-  {c[0]}+MMd/-------...-:sdds{c[1]} MMM   {r[11]}
-{c[1]}     -NMm-  {c[0]}:hNMNNNmdddddddddy/`{c[1]} dMM   {r[12]}
-{c[1]}      -dMNs-``{c[0]}-::::-------.``{c[1]}    dMM   {r[13]}
-{c[1]}       `/dMNmy+/:-------------:/yMMM   {r[14]}
-{c[1]}          ./ydNMMMMMMMMMMMMMMMMMMMMM   {r[15]}
-{c[1]}                                       {r[16]}
-{c[1]}                                       {r[17]}\n""",
-    Distributions.OPENSUSE: """
-{c[0]}            .;ldkO0000Okdl;.              {r[0]}
-{c[0]}        .;d00xl:^''''''^:ok00d;.          {r[1]}
-{c[0]}      .d00l'                'o00d.        {r[2]}
-{c[0]}    .d0Kd'  {c[1]}Okxol:;,.{c[0]}          :O0d.      {r[3]}
-{c[0]}   .OK{c[1]}KKK0kOKKKKKKKKKKOxo:,{c[0]}      lKO.     {r[4]}
-{c[0]}  ,0K{c[1]}KKKKKKKKKKKKKKK0P^{c[0]},,,{c[1]}^dx:{c[0]}    ;00,    {r[5]}
-{c[0]} .OK{c[1]}KKKKKKKKKKKKKKKk'{c[0]}.oOPPb.{c[1]}'0k.{c[0]}   cKO.   {r[6]}
-{c[0]} :KK{c[1]}KKKKKKKKKKKKKKK: {c[0]}kKx..dd{c[1]} lKd{c[0]}   'OK:   {r[7]}
-{c[0]} dKK{c[1]}KKKKKKKKKOx0KKKd {c[0]}^0KKKO'{c[1]} kKKc{c[0]}   dKd   {r[8]}
-{c[0]} dKK{c[1]}KKKKKKKKKK;.;oOKx,..{c[0]}^{c[1]}..;kKKK0.{c[0]}  dKd   {r[9]}
-{c[0]} :KK{c[1]}KKKKKKKKKK0o;...^cdxxOK0O/^^'{c[0]}  .0K:   {r[10]}
-{c[0]}  kKK{c[1]}KKKKKKKKKKKKK0x;,,......,;od{c[0]}  lKk    {r[11]}
-{c[0]}  '0K{c[1]}KKKKKKKKKKKKKKKKKKKK00KKOo^{c[0]}  c00'    {r[12]}
-{c[0]}   'kK{c[1]}KKOxddxkOO00000Okxoc;''{c[0]}   .dKk'     {r[13]}
-{c[0]}     l0Ko.                    .c00l'      {r[14]}
-{c[0]}      'l0Kk:.              .;xK0l'        {r[15]}
-{c[0]}         'lkK0xl:;,,,,;:ldO0kl'           {r[16]}
-{c[0]}             '^:ldxkkkkxdl:^'             {r[17]}\n""",
-    Distributions.RED_HAT: """
-{c[1]}                                             {r[0]}
-{c[1]}              {c[2]}\\`.-..........\\`{c[1]}               {r[1]}
-{c[1]}             {c[2]}\\`////////::.\\`-/.{c[1]}              {r[2]}
-{c[1]}             {c[2]}-: ....-////////.{c[1]}               {r[3]}
-{c[1]}             {c[2]}//:-::///////////\\`{c[1]}             {r[4]}
-{c[1]}      {c[2]}\\`--::: \\`-://////////////:{c[1]}            {r[5]}
-{c[1]}      {c[2]}//////-    \\`\\`.-://///////{c[1]} .\\`        {r[6]}
-{c[1]}      {c[2]}\\`://////:-.\\`    :///////::///:\\`{c[1]}     {r[7]}
-{c[1]}        {c[2]}.-/////////:---/////////////:{c[1]}        {r[8]}
-{c[1]}           {c[2]}.-://////////////////////.{c[1]}        {r[9]}
-{c[1]}          {c[0]}yMN+\\`.-${c[2]}::///////////////-\\`{c[1]}      {r[10]}
-{c[1]}       {c[0]}.-\\`:NMMNMs\\`  \\`..-------..\\`{c[1]}        {r[11]}
-{c[1]}        {c[0]}MN+/mMMMMMhoooyysshsss{c[1]}               {r[12]}
-{c[1]} {c[0]}MMM    MMMMMMMMMMMMMMyyddMMM+{c[1]}               {r[13]}
-{c[1]}  {c[0]}MMMM   MMMMMMMMMMMMMNdyNMMh\\`     hyhMMM{c[1]}   {r[14]}
-{c[1]}   {c[0]}MMMMMMMMMMMMMMMMyoNNNMMM+.   MMMMMMMM{c[1]}     {r[15]}
-{c[1]}    {c[0]}MMNMMMNNMMMMMNM+ mhsMNyyyyMNMMMMsMM{c[1]}      {r[16]}
-{c[1]}                                             {r[17]}\n""",
-    Distributions.SLACKWARE: """
-{c[0]}                  {c[1]}:::::::{c[0]}                       \0
-{c[0]}             {c[1]}:::::::::::::::::::{c[0]}                \0
-{c[0]}          {c[1]}:::::::::::::::::::::::::{c[0]}             {r[0]}
-{c[0]}        {c[1]}::::::::{c[2]}cllcccccllllllll{c[1]}::::::{c[0]}          {r[1]}
-{c[0]}     {c[1]}:::::::::{c[2]}lc{c[0]}               {c[2]}dc{c[1]}:::::::{c[0]}        {r[2]}
-{c[0]}    {c[1]}::::::::{c[2]}cl{c[0]}   {c[2]}clllccllll{c[0]}    {c[2]}oc{c[1]}:::::::::{c[0]}      {r[3]}
-{c[0]}   {c[1]}:::::::::{c[2]}o{c[0]}   {c[2]}lc{c[1]}::::::::{c[2]}co{c[0]}   {c[2]}oc{c[1]}::::::::::{c[0]}     {r[4]}
-{c[0]}  {c[1]}::::::::::{c[2]}o{c[0]}    {c[2]}cccclc{c[1]}:::::{c[2]}clcc{c[1]}::::::::::::{c[0]}    {r[5]}
-{c[0]}  {c[1]}:::::::::::{c[2]}lc{c[0]}        {c[2]}cclccclc{c[1]}:::::::::::::{c[0]}    {r[6]}
-{c[0]} {c[1]}::::::::::::::{c[2]}lcclcc{c[0]}          {c[2]}lc{c[1]}::::::::::::{c[0]}   {r[7]}
-{c[0]} {c[1]}::::::::::{c[2]}cclcc{c[1]}:::::{c[2]}lccclc{c[0]}     {c[2]}oc{c[1]}:::::::::::{c[0]}   {r[8]}
-{c[0]} {c[1]}::::::::::{c[2]}o{c[0]}    {c[2]}l{c[1]}::::::::::{c[2]}l{c[0]}    {c[2]}lc{c[1]}:::::::::::{c[0]}   {r[9]}
-{c[0]}  {c[1]}:::::{c[2]}c{c[0]} {c[0]}{c[1]}::{c[2]}o{c[0]}     {c[2]}clcllcccll{c[0]}     {c[2]}o{c[1]}:::::::::::{c[0]}    {r[10]}
-{c[0]}  {c[1]}:::::{c[2]}o{c[0]} {c[0]}{c[1]}:{c[2]}o{c[0]}                  {c[2]}clc{c[1]}:::::::::::{c[0]}     {r[11]}
-{c[0]}   {c[1]}::::{c[2]}o{c[0]} {c[0]}{c[1]}:{c[2]}ccslclccclclccclclc{c[1]}:::::::::::::{c[0]}      {r[12]}
-{c[0]}    {c[1]}:::{c[2]}o{c[0]}                             {c[1]}:::::{c[0]}      {r[13]}
-{c[0]}     {c[1]}::{c[2]}lcccccccccccccccccccccccccccco{c[1]}::::{c[0]}       {r[14]}
-{c[0]}       {c[1]}::::::::::::::::::::::::::::::::{c[0]}         {r[15]}
-{c[0]}         {c[1]}::::::::::::::::::::::::::::{c[0]}           {r[16]}
-{c[0]}            {c[1]}::::::::::::::::::::::{c[0]}              {r[17]}
-{c[0]}                 {c[1]}::::::::::::{c[0]}                   \n""",
-    Distributions.UBUNTU: """
-{c[1]}                           {c[0]}.oyhhs:{c[1]}     {r[0]}
-{c[1]}                  ..--.., {c[0]}shhhhhh-{c[1]}     {r[1]}
-{c[1]}                -+++++++++`:{c[0]}yyhhyo`{c[1]}    {r[2]}
-{c[1]}           {c[2]}.--{c[1]}  -++++++++/-.-{c[0]}::-`{c[1]}      {r[3]}
-{c[1]}         {c[2]}.::::-{c[1]}   :-----:/+++/++/.     {r[4]}
-{c[1]}        {c[2]}-:::::-.{c[1]}          .:++++++:    {r[5]}
-{c[1]}   ,,, {c[2]}.:::::-`{c[1]}             .++++++-   {r[6]}
-{c[1]} ./+++/-{c[2]}`-::-{c[1]}                ./////:   {r[7]}
-{c[1]} +++++++ {c[2]}.::-{c[1]}                          {r[8]}
-{c[1]} ./+++/-`{c[2]}-::-{c[1]}                {c[0]}:yyyyyo{c[1]}   {r[9]}
-{c[1]}   ``` `{c[2]}-::::-`{c[1]}             {c[0]}:yhhhhh:{c[1]}   {r[10]}
-{c[1]}       {c[2]}-:::::-.{c[1]}          {c[0]}`-ohhhhhh+{c[1]}    {r[11]}
-{c[1]}         {c[2]}.::::-`{c[1]} {c[0]}-o+///+oyhhyyyhy:{c[1]}     {r[12]}
-{c[1]}          {c[2]}`.--{c[1]}  {c[0]}/yhhhhhhhy+{c[2]},....{c[1]}       {r[13]}
-{c[1]}                {c[0]}/hhhhhhhhh{c[2]}-.-:::;{c[1]}      {r[14]}
-{c[1]}                {c[0]}`.:://::- {c[2]}-:::::;{c[1]}      {r[15]}
-{c[1]}                          {c[2]}`.-:-'{c[1]}       {r[16]}
-{c[1]}                                       {r[17]}\n""",
-    Distributions.WINDOWS: """
-{c[1]}                                        {r[0]}
-{c[1]}         {c[0]},.=:^!^!t3Z3z.,{c[1]}                {r[1]}
-{c[1]}        {c[0]}:tt:::tt333EE3{c[1]}                  {r[2]}
-{c[1]}        {c[0]}Et:::ztt33EEE{c[1]}  {c[2]}@Ee.,{c[1]}      {c[2]}..,{c[1]}   {r[3]}
-{c[1]}       {c[0]};tt:::tt333EE7{c[1]} {c[2]};EEEEEEttttt33#{c[1]}   {r[4]}
-{c[1]}      {c[0]}:Et:::zt333EEQ.{c[1]} {c[2]}SEEEEEttttt33QL{c[1]}   {r[5]}
-{c[1]}      {c[0]}it::::tt333EEF{c[1]} {c[2]}@EEEEEEttttt33F{c[1]}    {r[6]}
-{c[1]}     {c[0]};3=*^```'*4EEV{c[1]} {c[2]}:EEEEEEttttt33@.{c[1]}    {r[7]}
-{c[1]}     ,.=::::it=.,{c[0]} `{c[1]} {c[2]}@EEEEEEtttz33QF{c[1]}     {r[8]}
-{c[1]}    ;::::::::zt33){c[1]}   {c[2]}'4EEEtttji3P*{c[1]}      {r[9]}
-{c[1]}   :t::::::::tt33.{c[3]}:Z3z..{c[2]}  `` {c[3]},..g.{c[1]}      {r[10]}
-{c[1]}   i::::::::zt33F{c[1]} {c[3]}AEEEtttt::::ztF{c[1]}       {r[11]}
-{c[1]}  ;:::::::::t33V{c[1]} {c[3]};EEEttttt::::t3{c[1]}        {r[12]}
-{c[1]}  E::::::::zt33L{c[1]} {c[3]}@EEEtttt::::z3F{c[1]}        {r[13]}
-{c[1]} {{3=*^```'*4E3){c[1]} {c[3]};EEEtttt:::::tZ`{c[1]}        {r[14]}
-{c[1]}             `{c[1]} {c[3]}:EEEEtttt::::z7{c[1]}          {r[15]}
-{c[1]}                 {c[3]}'VEzjt:;;z>*`{c[1]}          {r[16]}
-{c[1]}                                        {r[17]}\n"""
-}
-
+from distributions import Distributions
+from constants import (
+    COLOR_DICT,
+    DE_DICT,
+    WM_DICT,
+    LOGOS_DICT,
+)
 
 # ----------- Configuration ----------- #
