+++ archey/logos/__init__.py
@@ -1,0 +1,15 @@
+from .arch_linux import ARCH_LINUX
+from .bunsenlabs import BUNSENLABS
+from .crunchbang import CRUNCHBANG
+from .debian import DEBIAN
+from .fedora import FEDORA
+from .gentoo import GENTOO
+from .kali_linux import KALI_LINUX
+from .manjaro import MANJARO
+from .linux import LINUX
+from .linux_mint import LINUX_MINT
+from .openuse import OPENUSE
+from .red_hat import RED_HAT
+from .slackware import SLACKWARE
+from .ubuntu import UBUNTU
+from .windows import WINDOWS
