+++ archey/logos/bunsenlabs.py
@@ -1,0 +1,24 @@
+BUNSENLABS = """
+{c[0]}           .{c[1]}..{c[0]}+hhy+-`         \0
+{c[0]}        `+hd{c[1]}-{c[0]}+dddd{c[2]}hyso{c[0]}+//:    \0
+{c[0]}      `+dddd{c[1]}:-{c[0]}sdh/.           {r[0]}
+{c[0]}     -hdddddh{c[1]}-.{c[2]}/:{c[0]}             {r[1]}
+{c[0]}    /ddddddddd{c[1]}:```{c[0]}            {r[2]}
+{c[0]}   :ddddddddddd/              {r[3]}
+{c[0]}  `hdddddddddddd+             {r[4]}
+{c[0]}  /dddddddddddddd:            {r[5]}
+{c[0]}  odddds..sddddddh            {r[6]}
+{c[0]}  oddd/    /dddddd:           {r[7]}
+{c[0]}  +dd+      +ddddd+           {r[8]}
+{c[0]}  .dd`      `ddddd+           {r[9]}
+{c[0]}   oh        ydddd:           {r[10]}
+{c[0]}   `o        sdddh`           {r[11]}
+{c[0]}             yddd:            {r[12]}
+{c[0]}            `dddo             {r[13]}
+{c[0]}     :s     :dds              {r[14]}
+{c[0]}     yd/    yd+               {r[15]}
+{c[0]}   `sddy   :h-                {r[16]}
+{c[0]}  `sddys`  :`                 {r[17]}
+{c[0]} -hdy+`y+yo./+/:-             \0
+{c[0]} ...  .o++oso+/               \n
+"""
