+++ archey/logos/crunchbang.py
@@ -1,0 +1,21 @@
+CRUNCHBANG = """
+{c[0]}                 ___       ___      _    {r[0]}
+{c[0]}                /  /      /  /     | |   {r[1]}
+{c[0]}               /  /      /  /      | |   {r[2]}
+{c[0]}              /  /      /  /       | |   {r[3]}
+{c[0]}      _______/  /______/  /______  | |   {r[4]}
+{c[0]}     /______   _______   _______/  | |   {r[5]}
+{c[0]}           /  /      /  /          | |   {r[6]}
+{c[0]}          /  /      /  /           | |   {r[7]}
+{c[0]}         /  /      /  /            | |   {r[8]}
+{c[0]}  ______/  /______/  /______       | |   {r[9]}
+{c[0]} /_____   _______   _______/       | |   {r[10]}
+{c[0]}      /  /      /  /               | |   {r[11]}
+{c[0]}     /  /      /  /                |_|   {r[12]}
+{c[0]}    /  /      /  /                  _    {r[13]}
+{c[0]}   /  /      /  /                  | |   {r[14]}
+{c[0]}  /__/      /__/                   |_|   {r[15]}
+{c[0]}                                         {r[16]}
+{c[0]}                                         {r[17]}\n
+""",
+
