+++ archey/logos/debian.py
@@ -1,0 +1,20 @@
+DEBIAN = """
+{c[1]}                                  {r[0]}
+{c[1]}          _sudZUZ#Z#XZo=_         {r[1]}
+{c[1]}       _jmZZ2!!~---~!!X##wx       {r[2]}
+{c[1]}    .<wdP~~            -!YZL,     {r[3]}
+{c[1]}   .mX2'       _xaaa__     XZ[.   {r[4]}
+{c[1]}   oZ[      _jdXY!~?S#wa   ]Xb;   {r[5]}
+{c[1]}  _#e'     .]X2(     ~Xw|  )XXc   {r[6]}
+{c[1]} .2Z`      ]X[.       xY|  ]oZ(   {r[7]}
+{c[1]} .2#;      )3k;     _s!~   jXf`   {r[8]}
+{c[1]}  {c[0]}1Z>{c[1]}      {c[0]}-]Xb/{c[1]}    {c[0]}~{c[1]}    {c[0]}__#2({c[1]}    {r[9]}
+{c[1]}  {c[0]}-Zo;{c[1]}       {c[0]}+!4ZwerfgnZZXY'{c[1]}      {r[10]}
+{c[1]}   {c[0]}*#[,{c[1]}        {c[0]}~-?!!!!!!-~{c[1]}        {r[11]}
+{c[1]}    {c[0]}XUb;.{c[1]}                         {r[12]}
+{c[1]}     {c[0]})YXL,,{c[1]}                       {r[13]}
+{c[1]}       {c[0]}+3#bc,{c[1]}                     {r[14]}
+{c[1]}         {c[0]}-)SSL,,{c[1]}                  {r[15]}
+{c[1]}            {c[0]}~~~~~{c[1]}                 {r[16]}
+{c[1]}                                  {r[17]}\n
+"""
