+++ archey/logos/fedora.py
@@ -1,0 +1,20 @@
+FEDORA = """
+{c[1]}           :/------------://          {r[0]}
+{c[1]}        :------------------://        {r[1]}
+{c[1]}      :-----------{c[0]}/shhdhyo/{c[1]}-://       {r[2]}
+{c[1]}    /-----------{c[0]}omMMMNNNMMMd/{c[1]}-:/      {r[3]}
+{c[1]}   :-----------{c[0]}sMMMdo:/{c[1]}       -:/     {r[4]}
+{c[1]}  :-----------{c[0]}:MMMd{c[1]}-------    --:/    {r[5]}
+{c[1]}  /-----------{c[0]}:MMMy{c[1]}-------    ---/    {r[6]}
+{c[1]} :------    --{c[0]}/+MMMh/{c[1]}--        ---:   {r[7]}
+{c[1]} :---     {c[0]}oNMMMMMMMMMNho{c[1]}     -----:   {r[8]}
+{c[1]} :--      {c[0]}+shhhMMMmhhy++{c[1]}   ------:    {r[9]}
+{c[1]} :-      -----{c[0]}:MMMy{c[1]}--------------/    {r[10]}
+{c[1]} :-     ------{c[0]}/MMMy{c[1]}-------------:     {r[11]}
+{c[1]} :-      ----{c[0]}/hMMM+{c[1]}------------:      {r[12]}
+{c[1]} :--{c[0]}:dMMNdhhdNMMNo{c[1]}-----------:        {r[13]}
+{c[1]} :---{c[0]}:sdNMMMMNds:{c[1]}----------:          {r[14]}
+{c[1]} :------{c[0]}:://:{c[1]}-----------://           {r[15]}
+{c[1]} :--------------------://             {r[16]}
+{c[1]}                                      {r[17]}\n
+"""
