+++ archey/logos/kali_linux.py
@@ -1,0 +1,24 @@
+KALI_LINUX = """
+{c[0]}      ,.....                                        \0
+{c[0]}  ----`     `..,;:ccc,.                             \0
+{c[0]}           ......''';lxO.                           {r[0]}
+{c[0]} .....''''..........,:ld;                           {r[1]}
+{c[0]}            .';;;:::;,,.x,                          {r[2]}
+{c[0]}       ..'''.            0Xxoc:,.  ...              {r[3]}
+{c[0]}   ....                ,ONkc;,;cokOdc',.            {r[4]}
+{c[0]}  .                   OMo           ':{c[1]}d{c[0]}o.           {r[5]}
+{c[0]}                     dMc               :OO;         {r[6]}
+{c[0]}                     0M.                 .:o.       {r[7]}
+{c[0]}                     ;Wd                            {r[8]}
+{c[0]}                      ;XO,                          {r[9]}
+{c[0]}                        ,d0Odlc;,..                 {r[10]}
+{c[0]}                            ..',;:cdOOd::,.         {r[11]}
+{c[0]}                                     .:d;.':;.      {r[12]}
+{c[0]}                                        'd,  .'     {r[13]}
+{c[0]}                                          ;l   ..   {r[14]}
+{c[0]}                                           .o       {r[15]}
+{c[0]}                                             c      {r[16]}
+{c[0]}                                             .'     {r[17]}
+{c[0]}                                              .     \n
+"""
+
