+++ archey/logos/linux.py
@@ -1,0 +1,21 @@
+LINUX = """
+{c[1]}                           {r[0]}
+{c[1]}          a8888b.          {r[1]}
+{c[1]}         d888888b.         {r[2]}
+{c[1]}         8P"YP"Y88         {r[3]}
+{c[1]}         8|o||o|88         {r[4]}
+{c[1]}         8{c[0]}\\vvvv/{c[1]}88         {r[5]}
+{c[1]}         8{c[0]} \\vv/ {c[1]}Y8.        {r[6]}
+{c[1]}        d/  {c[0]}`'{c[1]}  \\8b.       {r[7]}
+{c[1]}      .dP   .     Y8b.     {r[8]}
+{c[1]}     d8:'   "   `::88b.    {r[9]}
+{c[1]}    d8"           `Y88b    {r[10]}
+{c[1]}   :8P     '       :888    {r[11]}
+{c[1]}    8a.    :      _a88P    {r[12]}
+{c[1]}  {c[0]}._/"{c[1]}Yaa_ :    .{c[0]}| {c[1]}88P{c[0]}|{c[1]}    {r[13]}
+{c[1]} {c[0]}\\++++{c[1]}YP"      `{c[0]}| {c[1]}8P{c[0]}++\\.{c[1]}   {r[14]}
+{c[1]} {c[0]}/+++++\\.{c[1]}_____.d{c[0]}|+++++/{c[1]}    {r[15]}
+{c[1]}  {c[0]}\\++++++){c[1]}888888P{c[0]}\\+++/{c[1]}     {r[16]}
+{c[1]}                           {r[17]}\n
+"""
+
