+++ archey/logos/linux_mint.py
@@ -1,0 +1,21 @@
+LINUX_MINT = """
+{c[1]}                                       {r[0]}
+{c[1]} MMMMMMMMMMMMMMMMMMMMMMMMMmds+.        {r[1]}
+{c[1]} MMm----::-://////////////oymNMd+`     {r[2]}
+{c[1]} MMd      {c[0]}/++{c[1]}                -sNMd:    {r[3]}
+{c[1]} MMNso/`  {c[0]}dMM{c[1]}    {c[0]}`.::-. .-::.`{c[1]} .hMN:   {r[4]}
+{c[1]} ddddMMh  {c[0]}dMM{c[1]}   {c[0]}:hNMNMNhNMNMNh:`{c[1]} NMm   {r[5]}
+{c[1]}     NMm  {c[0]}dMM{c[1]}  {c[0]}.NMN/-+MMM+-/NMN`{c[1]} dMM   {r[6]}
+{c[1]}     NMm  {c[0]}dMM{c[1]}  {c[0]}-MMm{c[1]}  {c[0]}`MMM{c[1]}   {c[0]}dMM.{c[1]} dMM   {r[7]}
+{c[1]}     NMm  {c[0]}dMM{c[1]}  {c[0]}-MMm{c[1]}  {c[0]}`MMM{c[1]}   {c[0]}dMM.{c[1]} dMM   {r[8]}
+{c[1]}     NMm  {c[0]}dMM{c[1]}  {c[0]}.mmd{c[1]}  {c[0]}`mmm{c[1]}   {c[0]}yMM.{c[1]} dMM   {r[9]}
+{c[1]}     NMm  {c[0]}dMM`{c[1]}  {c[0]}..`{c[1]}  {c[0]}`...{c[1]}   {c[0]}ydm.{c[1]} dMM   {r[10]}
+{c[1]}     hMM-  {c[0]}+MMd/-------...-:sdds{c[1]} MMM   {r[11]}
+{c[1]}     -NMm-  {c[0]}:hNMNNNmdddddddddy/`{c[1]} dMM   {r[12]}
+{c[1]}      -dMNs-``{c[0]}-::::-------.``{c[1]}    dMM   {r[13]}
+{c[1]}       `/dMNmy+/:-------------:/yMMM   {r[14]}
+{c[1]}          ./ydNMMMMMMMMMMMMMMMMMMMMM   {r[15]}
+{c[1]}                                       {r[16]}
+{c[1]}                                       {r[17]}\n
+"""
+
