+++ archey/logos/manjaro.py
@@ -1,0 +1,20 @@
+MANJARO = """
+{c[0]} $$$$$$$$$$$$$$$$  $$$$$$$   {r[0]}
+{c[0]} M77777777777777M  M77777M   {r[1]}
+{c[0]} M77777777777777M  M77777M   {r[2]}
+{c[0]} M77777MMMMMMMMMM  M77777M   {r[3]}
+{c[0]} M77777M           M77777M   {r[4]}
+{c[0]} M77777M  $$$$$$$  M77777M   {r[5]}
+{c[0]} MMMMMMM  M77777M  M77777M   {r[6]}
+{c[0]}          M77777M  M77777M   {r[7]}
+{c[0]} $$$$$$$  M77777M  M77777M   {r[8]}
+{c[0]} M77777M  M77777M  M77777M   {r[9]}
+{c[0]} M77777M  M77777M  M77777M   {r[10]}
+{c[0]} M77777M  M77777M  M77777M   {r[11]}
+{c[0]} M77777M  M77777M  M77777M   {r[12]}
+{c[0]} M77777M  M77777M  M77777M   {r[13]}
+{c[0]} M77777M  M77777M  M77777M   {r[14]}
+{c[0]} M77777M  M77777M  M77777M   {r[15]}
+{c[0]} M77777M  M77777M  M77777M   {r[16]}
+{c[0]} MMMMMMM  MMMMMMM  MMMMMMM   {r[17]}\n
+"""
