+++ archey/logos/red_hat.py
@@ -1,0 +1,20 @@
+RED_HAT = """
+{c[1]}                                             {r[0]}
+{c[1]}              {c[2]}\\`.-..........\\`{c[1]}               {r[1]}
+{c[1]}             {c[2]}\\`////////::.\\`-/.{c[1]}              {r[2]}
+{c[1]}             {c[2]}-: ....-////////.{c[1]}               {r[3]}
+{c[1]}             {c[2]}//:-::///////////\\`{c[1]}             {r[4]}
+{c[1]}      {c[2]}\\`--::: \\`-://////////////:{c[1]}            {r[5]}
+{c[1]}      {c[2]}//////-    \\`\\`.-://///////{c[1]} .\\`        {r[6]}
+{c[1]}      {c[2]}\\`://////:-.\\`    :///////::///:\\`{c[1]}     {r[7]}
+{c[1]}        {c[2]}.-/////////:---/////////////:{c[1]}        {r[8]}
+{c[1]}           {c[2]}.-://////////////////////.{c[1]}        {r[9]}
+{c[1]}          {c[0]}yMN+\\`.-${c[2]}::///////////////-\\`{c[1]}      {r[10]}
+{c[1]}       {c[0]}.-\\`:NMMNMs\\`  \\`..-------..\\`{c[1]}        {r[11]}
+{c[1]}        {c[0]}MN+/mMMMMMhoooyysshsss{c[1]}               {r[12]}
+{c[1]} {c[0]}MMM    MMMMMMMMMMMMMMyyddMMM+{c[1]}               {r[13]}
+{c[1]}  {c[0]}MMMM   MMMMMMMMMMMMMNdyNMMh\\`     hyhMMM{c[1]}   {r[14]}
+{c[1]}   {c[0]}MMMMMMMMMMMMMMMMyoNNNMMM+.   MMMMMMMM{c[1]}     {r[15]}
+{c[1]}    {c[0]}MMNMMMNNMMMMMNM+ mhsMNyyyyMNMMMMsMM{c[1]}      {r[16]}
+{c[1]}                                             {r[17]}\n
+"""
