+++ archey/logos/slackware.py
@@ -1,0 +1,23 @@
+SLACKWARE = """
+{c[0]}                  {c[1]}:::::::{c[0]}                       \0
+{c[0]}             {c[1]}:::::::::::::::::::{c[0]}                \0
+{c[0]}          {c[1]}:::::::::::::::::::::::::{c[0]}             {r[0]}
+{c[0]}        {c[1]}::::::::{c[2]}cllcccccllllllll{c[1]}::::::{c[0]}          {r[1]}
+{c[0]}     {c[1]}:::::::::{c[2]}lc{c[0]}               {c[2]}dc{c[1]}:::::::{c[0]}        {r[2]}
+{c[0]}    {c[1]}::::::::{c[2]}cl{c[0]}   {c[2]}clllccllll{c[0]}    {c[2]}oc{c[1]}:::::::::{c[0]}      {r[3]}
+{c[0]}   {c[1]}:::::::::{c[2]}o{c[0]}   {c[2]}lc{c[1]}::::::::{c[2]}co{c[0]}   {c[2]}oc{c[1]}::::::::::{c[0]}     {r[4]}
+{c[0]}  {c[1]}::::::::::{c[2]}o{c[0]}    {c[2]}cccclc{c[1]}:::::{c[2]}clcc{c[1]}::::::::::::{c[0]}    {r[5]}
+{c[0]}  {c[1]}:::::::::::{c[2]}lc{c[0]}        {c[2]}cclccclc{c[1]}:::::::::::::{c[0]}    {r[6]}
+{c[0]} {c[1]}::::::::::::::{c[2]}lcclcc{c[0]}          {c[2]}lc{c[1]}::::::::::::{c[0]}   {r[7]}
+{c[0]} {c[1]}::::::::::{c[2]}cclcc{c[1]}:::::{c[2]}lccclc{c[0]}     {c[2]}oc{c[1]}:::::::::::{c[0]}   {r[8]}
+{c[0]} {c[1]}::::::::::{c[2]}o{c[0]}    {c[2]}l{c[1]}::::::::::{c[2]}l{c[0]}    {c[2]}lc{c[1]}:::::::::::{c[0]}   {r[9]}
+{c[0]}  {c[1]}:::::{c[2]}c{c[0]} {c[0]}{c[1]}::{c[2]}o{c[0]}     {c[2]}clcllcccll{c[0]}     {c[2]}o{c[1]}:::::::::::{c[0]}    {r[10]}
+{c[0]}  {c[1]}:::::{c[2]}o{c[0]} {c[0]}{c[1]}:{c[2]}o{c[0]}                  {c[2]}clc{c[1]}:::::::::::{c[0]}     {r[11]}
+{c[0]}   {c[1]}::::{c[2]}o{c[0]} {c[0]}{c[1]}:{c[2]}ccslclccclclccclclc{c[1]}:::::::::::::{c[0]}      {r[12]}
+{c[0]}    {c[1]}:::{c[2]}o{c[0]}                             {c[1]}:::::{c[0]}      {r[13]}
+{c[0]}     {c[1]}::{c[2]}lcccccccccccccccccccccccccccco{c[1]}::::{c[0]}       {r[14]}
+{c[0]}       {c[1]}::::::::::::::::::::::::::::::::{c[0]}         {r[15]}
+{c[0]}         {c[1]}::::::::::::::::::::::::::::{c[0]}           {r[16]}
+{c[0]}            {c[1]}::::::::::::::::::::::{c[0]}              {r[17]}
+{c[0]}                 {c[1]}::::::::::::{c[0]}                   \n
+"""
