+++ archey/logos/ubuntu.py
@@ -1,0 +1,21 @@
+UBUNTU = """
+{c[1]}                           {c[0]}.oyhhs:{c[1]}     {r[0]}
+{c[1]}                  ..--.., {c[0]}shhhhhh-{c[1]}     {r[1]}
+{c[1]}                -+++++++++`:{c[0]}yyhhyo`{c[1]}    {r[2]}
+{c[1]}           {c[2]}.--{c[1]}  -++++++++/-.-{c[0]}::-`{c[1]}      {r[3]}
+{c[1]}         {c[2]}.::::-{c[1]}   :-----:/+++/++/.     {r[4]}
+{c[1]}        {c[2]}-:::::-.{c[1]}          .:++++++:    {r[5]}
+{c[1]}   ,,, {c[2]}.:::::-`{c[1]}             .++++++-   {r[6]}
+{c[1]} ./+++/-{c[2]}`-::-{c[1]}                ./////:   {r[7]}
+{c[1]} +++++++ {c[2]}.::-{c[1]}                          {r[8]}
+{c[1]} ./+++/-`{c[2]}-::-{c[1]}                {c[0]}:yyyyyo{c[1]}   {r[9]}
+{c[1]}   ``` `{c[2]}-::::-`{c[1]}             {c[0]}:yhhhhh:{c[1]}   {r[10]}
+{c[1]}       {c[2]}-:::::-.{c[1]}          {c[0]}`-ohhhhhh+{c[1]}    {r[11]}
+{c[1]}         {c[2]}.::::-`{c[1]} {c[0]}-o+///+oyhhyyyhy:{c[1]}     {r[12]}
+{c[1]}          {c[2]}`.--{c[1]}  {c[0]}/yhhhhhhhy+{c[2]},....{c[1]}       {r[13]}
+{c[1]}                {c[0]}/hhhhhhhhh{c[2]}-.-:::;{c[1]}      {r[14]}
+{c[1]}                {c[0]}`.:://::- {c[2]}-:::::;{c[1]}      {r[15]}
+{c[1]}                          {c[2]}`.-:-'{c[1]}       {r[16]}
+{c[1]}                                       {r[17]}\n
+"""
+
