+++ archey/logos/windows.py
@@ -1,0 +1,21 @@
+WINDOWS = """
+{c[1]}                                        {r[0]}
+{c[1]}         {c[0]},.=:^!^!t3Z3z.,{c[1]}                {r[1]}
+{c[1]}        {c[0]}:tt:::tt333EE3{c[1]}                  {r[2]}
+{c[1]}        {c[0]}Et:::ztt33EEE{c[1]}  {c[2]}@Ee.,{c[1]}      {c[2]}..,{c[1]}   {r[3]}
+{c[1]}       {c[0]};tt:::tt333EE7{c[1]} {c[2]};EEEEEEttttt33#{c[1]}   {r[4]}
+{c[1]}      {c[0]}:Et:::zt333EEQ.{c[1]} {c[2]}SEEEEEttttt33QL{c[1]}   {r[5]}
+{c[1]}      {c[0]}it::::tt333EEF{c[1]} {c[2]}@EEEEEEttttt33F{c[1]}    {r[6]}
+{c[1]}     {c[0]};3=*^```'*4EEV{c[1]} {c[2]}:EEEEEEttttt33@.{c[1]}    {r[7]}
+{c[1]}     ,.=::::it=.,{c[0]} `{c[1]} {c[2]}@EEEEEEtttz33QF{c[1]}     {r[8]}
+{c[1]}    ;::::::::zt33){c[1]}   {c[2]}'4EEEtttji3P*{c[1]}      {r[9]}
+{c[1]}   :t::::::::tt33.{c[3]}:Z3z..{c[2]}  `` {c[3]},..g.{c[1]}      {r[10]}
+{c[1]}   i::::::::zt33F{c[1]} {c[3]}AEEEtttt::::ztF{c[1]}       {r[11]}
+{c[1]}  ;:::::::::t33V{c[1]} {c[3]};EEEttttt::::t3{c[1]}        {r[12]}
+{c[1]}  E::::::::zt33L{c[1]} {c[3]}@EEEtttt::::z3F{c[1]}        {r[13]}
+{c[1]} {{3=*^```'*4E3){c[1]} {c[3]};EEEtttt:::::tZ`{c[1]}        {r[14]}
+{c[1]}             `{c[1]} {c[3]}:EEEEtttt::::z7{c[1]}          {r[15]}
+{c[1]}                 {c[3]}'VEzjt:;;z>*`{c[1]}          {r[16]}
+{c[1]}                                        {r[17]}\n
+"""
+
